Add get_proc_address to NativeGLContextMethods
NOTE: Not directly merged since it's not tested for macos
