+++ src/platform/mod.rs
@@ -2,9 +2,12 @@
 use layers::platform::surface::NativeGraphicsMetadata;
 
 pub trait NativeGLContextMethods {
+    fn get_proc_address(&str) -> *const ();
+
     fn create_headless() -> Result<Self, &'static str>;
     fn is_current(&self) -> bool;
     fn make_current(&self) -> Result<(), &'static str>;
+
     #[cfg(feature="texture_surface")]
     fn get_metadata(&self) -> NativeGraphicsMetadata;
