+++ src/platform/not_implemented/native_gl_context.rs
@@ -6,6 +6,10 @@ pub struct NativeGLContext;
 use layers::platform::surface::NativeGraphicsMetadata;
 
 impl NativeGLContextMethods for NativeGLContext {
+    fn get_proc_address(addr: &str) -> *const () {
+        0 as *const ()
+    }
+
     fn create_headless() -> Result<NativeGLContext, &'static str> {
         Err("Not implemented (yet)")
     }
