+++ src/tests.rs
@@ -5,22 +5,6 @@ use GLContext;
 use GLContextAttributes;
 use ColorAttachmentType;
 
-use std::ffi::CString;
-
-#[cfg(target_os = "linux")]
-use glx;
-#[cfg(target_os = "android")]
-use egl;
-
-#[cfg(target_os = "macos")]
-use core_foundation::bundle::{CFBundleGetBundleWithIdentifier, CFBundleGetFunctionPointerForName};
-#[cfg(target_os = "macos")]
-use core_foundation::base::TCFType;
-#[cfg(target_os = "macos")]
-use core_foundation::string::CFString;
-
-use std::str::FromStr;
-
 #[cfg(feature="texture_surface")]
 use layers::texturegl::Texture;
 
@@ -38,37 +22,6 @@ extern {}
 // This is probably a time bomb
 static mut GL_LOADED : bool = false;
 
-// Shamelessly stolen from glutin
-#[cfg(target_os = "linux")]
-fn get_proc_address(addr: &str) -> *const () {
-        let addr = CString::new(addr.as_bytes()).unwrap();
-        let addr = addr.as_ptr();
-        unsafe {
-            glx::GetProcAddress(addr as *const _) as *const ()
-        }
-}
-
-#[cfg(target_os = "android")]
-fn get_proc_address(addr: &str) -> *const () {
-        let addr = CString::new(s.as_bytes()).unwrap();
-        let addr = addr.as_ptr();
-        unsafe {
-            egl::GetProcAddress(addr as *const _) as *const ()
-        }
-}
-
-#[cfg(target_os = "macos")]
-fn get_proc_address(addr: &str) -> *const () {
-        let symbol_name: CFString = FromStr::from_str(addr).unwrap();
-        let framework_name: CFString = FromStr::from_str("com.apple.opengl").unwrap();
-        let framework = unsafe {
-            CFBundleGetBundleWithIdentifier(framework_name.as_concrete_TypeRef())
-        };
-        let symbol = unsafe {
-            CFBundleGetFunctionPointerForName(framework, symbol_name.as_concrete_TypeRef())
-        };
-        symbol as *const ()
-}
 
 fn load_gl() {
     unsafe {
@@ -76,7 +29,7 @@ fn load_gl() {
             return;
         }
 
-        gl::load_with(|s| get_proc_address(s) as *const _);
+        gl::load_with(|s| GLContext::get_proc_address(s) as *const _);
         GL_LOADED = true;
     }
 }
